+++ src/lib/Arc.svelte
@@ -22,9 +22,11 @@ Further information: [Konva API docs](https://konvajs.org/api/Konva.Arc.html), [
 	import Konva from 'konva';
 	import { onMount, onDestroy, createEventDispatcher } from 'svelte';
 	import type { Writable } from 'svelte/store';
-	import { registerEvents } from '$lib/util/events';
+	import { registerEvents, type KonvaEvents } from '$lib/util/events';
 	import { getParentContainer, type KonvaParent } from '$lib/util/manageContext';
 	import { copyExistingKeys } from '$lib/util/object';
+
+	interface $$Events extends KonvaEvents {}
 
 	export let config: Konva.ArcConfig;
 	export let handle = new Konva.Arc(config);