+++ src/lib/util/events.ts
@@ -37,6 +37,48 @@ const KONVA_EVENTS = [
 	'transformend'
 ];
 
+export type KonvaMouseEvent = CustomEvent<Konva.KonvaEventObject<MouseEvent>>;
+export type KonvaWheelEvent = CustomEvent<Konva.KonvaEventObject<WheelEvent>>;
+export type KonvaTouchEvent = CustomEvent<Konva.KonvaEventObject<TouchEvent>>;
+export type KonvaPointerEvent = CustomEvent<Konva.KonvaEventObject<PointerEvent>>;
+export type KonvaDragTransformEvent = CustomEvent<
+	Konva.KonvaEventObject<MouseEvent | PointerEvent | TouchEvent>
+>;
+
+export interface KonvaEvents {
+	mouseover: KonvaMouseEvent;
+	mouseout: KonvaMouseEvent;
+	mouseenter: KonvaMouseEvent;
+	mouseleave: KonvaMouseEvent;
+	mousemove: KonvaMouseEvent;
+	mousedown: KonvaMouseEvent;
+	mouseup: KonvaMouseEvent;
+	wheel: KonvaWheelEvent;
+	click: KonvaMouseEvent;
+	dblclick: KonvaMouseEvent;
+	touchstart: KonvaTouchEvent;
+	touchmove: KonvaTouchEvent;
+	touchend: KonvaTouchEvent;
+	tap: KonvaTouchEvent;
+	dbltap: KonvaTouchEvent;
+	pointerdown: KonvaPointerEvent;
+	pointermove: KonvaPointerEvent;
+	pointerup: KonvaPointerEvent;
+	pointercancel: KonvaPointerEvent;
+	pointerover: KonvaPointerEvent;
+	pointerenter: KonvaPointerEvent;
+	pointerout: KonvaPointerEvent;
+	pointerleave: KonvaPointerEvent;
+	pointerclick: KonvaPointerEvent;
+	pointerdblclick: KonvaPointerEvent;
+	dragstart: KonvaDragTransformEvent;
+	dragmove: KonvaDragTransformEvent;
+	dragend: KonvaDragTransformEvent;
+	transformstart: KonvaDragTransformEvent;
+	transform: KonvaDragTransformEvent;
+	transformend: KonvaDragTransformEvent;
+}
+
 /**
  * Registers all possible Konva node events with the provided Svelte dispatcher
  *
@@ -44,7 +86,9 @@ const KONVA_EVENTS = [
  * @param node
  */
 export function registerEvents(
-	dispatch: EventDispatcher<any>, // TODO: Check if the Konva payloads can be fully typed
+	dispatch: EventDispatcher<
+		Record<string, Konva.KonvaEventObject<MouseEvent | PointerEvent | TouchEvent>>
+	>,
 	node: Konva.Node
 ) {
 	KONVA_EVENTS.forEach((event) => {