+++ CHANGELOG.md
@@ -1,7 +1,9 @@
 # Unreleased
 
+- Added component event typings. This should significantly improve autocompletion and suggestions when using the on:event syntax for Konva events
+- Added strong typings for component event payloads. This change might break your typings in case you used a wrong type for the event payload in your event handlers. To fix this, type the event payload of your function using the new payload types exported by svelte-konva: `KonvaMouseEvent`, `KonvaWheelEvent`, `KonvaTouchEvent`, `KonvaPointerEvent` and `KonvaDragTransformEvent`
 - Tighten peer dependencies to only match specific major versions instead of matching all major versions starting from a certain version. This avoids svelte-konva being used in future major versions of Svelte or Konva which might break it
-- Updated to Svelte 4
+- Updated to Svelte 4 (the library is still fully compatible with Svelte v3)
 - Change Testing from Jest to Vitest which simplifies the configuration a lot
 - Various dependency updates